Move session repository
